Add files via upload
The algorithms used in this project is Convolutional Neural Networks (CNN) and Random forest (RF) are presented. The results showed that when compared with other classifiers , and RF performed well in terms of detection accuracy. Machine learning is widely used in this field by antivirus and anti malware programs as well as by these malicious programs. The next segment deals with the design of the proposed model and its working. The result of our project is to detect the malware in APK file using Convolutional Neural Networks (CNN) and Random forest (RF) Algorithm.